Replace version display handling Original rev  : rc@rc0.org.uk--2003-spill/spill--main--0--patch-9
Previously the code relied on the $Name$ construct which CVS filled in when
exporting the source tree given a named tag.  This method isn't going to work
any more in an arch-based world.  Create a new version.h file that contains the
version string.  This will get overwritten with a string indicating the correct
version as part of the make-release process.